--- server/db_mongodb.go
@@ -1,3 +1,0 @@
-package server
-
-type mongodb struct{}
